+++ src/core/asset_manager.h
@@ -32,7 +32,13 @@ namespace Aegix
 		/// @return Texture with the data from the file
 		std::shared_ptr<Graphics::Texture> createTexture(const std::filesystem::path& texturePath, const Graphics::Texture::Config& config = {});
 
-		std::shared_ptr<Graphics::Texture> createTexture(const glm::vec4& color, size_t width = 1, size_t height = 1, const Graphics::Texture::Config& config = {});
+		/// @brief Creates a texture with a single color
+		/// @param color Color of the texture
+		/// @param width Width of the texture
+		/// @param height Height of the texture
+		/// @param config Configuration of the texture
+		/// @return Created texture 
+		std::shared_ptr<Graphics::Texture> createTexture(const glm::vec4& color, uint32_t width = 1, uint32_t height = 1, const Graphics::Texture::Config& config = {});
 
 		/// @brief Adds a render system to the renderer and returns a reference to it
 		/// @tparam T Type of the render system to add
